fix(contacts): update behaviour when adding contact
